+++ internal/framework/service/rulesets/resource.go
@@ -77,16 +77,14 @@ func (r *RulesetResource) Create(ctx context.Context, req resource.CreateRequest
 	rulesetPhase := data.Phase.ValueString()
 	rulesetShareableEntitlementName := data.ShareableEntitlementName.ValueString()
 
-	var (
-		ruleset      cloudflare.Ruleset
-		semaphoreErr error
-	)
-
+	var identifier *cloudflare.ResourceContainer
 	if accountID.ValueString() != "" {
-		ruleset, semaphoreErr = r.client.GetAccountRulesetPhase(ctx, accountID.ValueString(), rulesetPhase)
+		identifier = cloudflare.AccountIdentifier(accountID.ValueString())
 	} else {
-		ruleset, semaphoreErr = r.client.GetZoneRulesetPhase(ctx, zoneID.ValueString(), rulesetPhase)
+		identifier = cloudflare.ZoneIdentifier(zoneID.ValueString())
 	}
+
+	ruleset, semaphoreErr := r.client.GetEntrypointRuleset(ctx, identifier, rulesetPhase)
 
 	// If an entrypoint ruleset with the same kind already exists, we should
 	// prevent the user from accidentally overriding their existing
@@ -113,12 +111,8 @@ func (r *RulesetResource) Create(ctx context.Context, req resource.CreateRequest
 
 		if semaphoreErr == nil && len(ruleset.Rules) == 0 && ruleset.Description == "" {
 			tflog.Debug(ctx, "default entrypoint ruleset created by the UI with empty rules found, recreating from scratch")
-			var err error
-			if accountID.ValueString() != "" {
-				err = r.client.DeleteAccountRuleset(ctx, accountID.ValueString(), ruleset.ID)
-			} else {
-				err = r.client.DeleteZoneRuleset(ctx, zoneID.ValueString(), ruleset.ID)
-			}
+
+			err := r.client.DeleteRuleset(ctx, identifier, ruleset.ID)
 			if err != nil {
 				resp.Diagnostics.AddError(
 					fmt.Sprintf("failed to delete existing entrypoint ruleset with ID %q", ruleset.ID),
@@ -129,7 +123,7 @@ func (r *RulesetResource) Create(ctx context.Context, req resource.CreateRequest
 		}
 	}
 
-	rs := cloudflare.Ruleset{
+	rs := cloudflare.CreateRulesetParams{
 		Name:                     rulesetName,
 		Description:              rulesetDescription,
 		Kind:                     rulesetKind,
@@ -143,19 +137,15 @@ func (r *RulesetResource) Create(ctx context.Context, req resource.CreateRequest
 		rs.Rules = rulesetData.Rules
 	}
 
-	var rulesetCreateErr error
-	if accountID.ValueString() != "" {
-		ruleset, rulesetCreateErr = r.client.CreateAccountRuleset(ctx, accountID.ValueString(), rs)
-	} else {
-		ruleset, rulesetCreateErr = r.client.CreateZoneRuleset(ctx, zoneID.ValueString(), rs)
-	}
+	ruleset, rulesetCreateErr := r.client.CreateRuleset(ctx, identifier, rs)
 
 	if rulesetCreateErr != nil {
 		resp.Diagnostics.AddError(fmt.Sprintf("error creating ruleset %s", rulesetName), rulesetCreateErr.Error())
 		return
 	}
 
-	rulesetEntryPoint := cloudflare.Ruleset{
+	params := cloudflare.UpdateEntrypointRulesetParams{
+		Phase:       rulesetPhase,
 		Description: rulesetDescription,
 		Rules:       rulesetData.Rules,
 	}
@@ -164,12 +154,7 @@ func (r *RulesetResource) Create(ctx context.Context, req resource.CreateRequest
 	// For "custom" rulesets, we don't send a follow up PUT it to the entrypoint
 	// endpoint.
 	if rulesetKind != string(cloudflare.RulesetKindCustom) {
-		if accountID.ValueString() != "" {
-			ruleset, err = r.client.UpdateAccountRulesetPhase(ctx, accountID.ValueString(), rulesetPhase, rulesetEntryPoint)
-		} else {
-			ruleset, err = r.client.UpdateZoneRulesetPhase(ctx, zoneID.ValueString(), rulesetPhase, rulesetEntryPoint)
-		}
-
+		ruleset, err = r.client.UpdateEntrypointRuleset(ctx, identifier, params)
 		if err != nil {
 			resp.Diagnostics.AddError(fmt.Sprintf("error updating ruleset phase entrypoint %s", rulesetName), err.Error())
 			return
@@ -198,15 +183,15 @@ func (r *RulesetResource) Read(ctx context.Context, req resource.ReadRequest, re
 
 	accountID := data.AccountID
 	zoneID := data.ZoneID
-	var err error
-	var ruleset cloudflare.Ruleset
 
+	var identifier *cloudflare.ResourceContainer
 	if accountID.ValueString() != "" {
-		ruleset, err = r.client.GetAccountRuleset(ctx, accountID.ValueString(), data.ID.ValueString())
+		identifier = cloudflare.AccountIdentifier(accountID.ValueString())
 	} else {
-		ruleset, err = r.client.GetZoneRuleset(ctx, zoneID.ValueString(), data.ID.ValueString())
+		identifier = cloudflare.ZoneIdentifier(zoneID.ValueString())
 	}
 
+	ruleset, err := r.client.GetRuleset(ctx, identifier, data.ID.ValueString())
 	if err != nil {
 		var notFoundError *cloudflare.NotFoundError
 		if errors.As(err, &notFoundError) {
@@ -247,15 +232,19 @@ func (r *RulesetResource) Update(ctx context.Context, req resource.UpdateRequest
 		return
 	}
 
-	var err error
-	var rs cloudflare.Ruleset
-	description := plan.Description.ValueString()
+	var identifier *cloudflare.ResourceContainer
 	if accountID.ValueString() != "" {
-		rs, err = r.client.UpdateAccountRuleset(ctx, accountID.ValueString(), state.ID.ValueString(), description, remappedRules)
+		identifier = cloudflare.AccountIdentifier(accountID.ValueString())
 	} else {
-		rs, err = r.client.UpdateZoneRuleset(ctx, zoneID, state.ID.ValueString(), description, remappedRules)
+		identifier = cloudflare.ZoneIdentifier(zoneID)
 	}
 
+	params := cloudflare.UpdateRulesetParams{
+		ID:          state.ID.ValueString(),
+		Description: plan.Description.ValueString(),
+		Rules:       remappedRules,
+	}
+	rs, err := r.client.UpdateRuleset(ctx, identifier, params)
 	if err != nil {
 		resp.Diagnostics.AddError(fmt.Sprintf("error updating ruleset with ID %q", state.ID.ValueString()), err.Error())
 		return
@@ -275,14 +264,14 @@ func (r *RulesetResource) Delete(ctx context.Context, req resource.DeleteRequest
 	accountID := data.AccountID
 	zoneID := data.ZoneID
 
-	var err error
-
+	var identifier *cloudflare.ResourceContainer
 	if accountID.ValueString() != "" {
-		err = r.client.DeleteAccountRuleset(ctx, accountID.ValueString(), data.ID.ValueString())
+		identifier = cloudflare.AccountIdentifier(accountID.ValueString())
 	} else {
-		err = r.client.DeleteZoneRuleset(ctx, zoneID.ValueString(), data.ID.ValueString())
+		identifier = cloudflare.ZoneIdentifier(zoneID.ValueString())
 	}
 
+	err := r.client.DeleteRuleset(ctx, identifier, data.ID.ValueString())
 	if err != nil {
 		resp.Diagnostics.AddError(fmt.Sprintf("error deleting ruleset with ID %q", data.ID.ValueString()), err.Error())
 		return