+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 DEPENDENCIES:
 
+* provider: bumps github.com/hashicorp/terraform-plugin-framework from 1.3.2 to 1.3.3 ([#2627](https://github.com/cloudflare/terraform-provider-cloudflare/issues/2627))
 * provider: bumps github.com/hashicorp/terraform-plugin-mux from 0.11.1 to 0.11.2 ([#2616](https://github.com/cloudflare/terraform-provider-cloudflare/issues/2616))
 
 ## 4.10.0 (12th July, 2023)